refactors the names in log reporter
