--- layers/log_reporter/CHANGELOG.md
@@ -1,3 +1,0 @@
-## 0.0.1
-
-* TODO: Describe initial release.
