--- layers/log_reporter/LICENSE
@@ -1,1 +1,0 @@
-TODO: Add your license here.
