+++ layers/log_reporter/analysis_options.yaml
@@ -1,4 +1,1 @@
-include: package:flutter_lints/flutter.yaml
-
-# Additional information about this file can be found at
-# https://dart.dev/guides/language/analysis-options
+include: package:lints/recommended.yaml
