+++ layers/log_reporter/lib/log_reporter.dart
@@ -1,7 +1,3 @@
 library log_reporter;
 
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
+export 'src/main.dart';
