+++ layers/log_reporter/lib/src/di/di.dart
@@ -1,0 +1,3 @@
+import 'package:get_it/get_it.dart';
+
+void setupPackageDependencies(final GetIt getIt) {}
