--- layers/log_reporter/lib/src/firebase_log_reporter.dart
@@ -1,25 +1,0 @@
-import 'package:log_reporter/src/log_reporter.dart';
-
-class FirebaseLogReporter implements LogReporter {
-  @override
-  Future<void> debug(
-    String message, {
-    String? tag,
-    error,
-    StackTrace? stacktrace,
-  }) {
-    // TODO: implement debug
-    throw UnimplementedError();
-  }
-
-  @override
-  Future<void> error(
-    String message, {
-    String? tag,
-    error,
-    StackTrace? stacktrace,
-  }) {
-    // TODO: implement error
-    throw UnimplementedError();
-  }
-}
