+++ layers/log_reporter/lib/src/log/log_reporter.dart
@@ -1,15 +1,15 @@
 abstract class LogReporter {
-  Future<void> debug(
+  void debug(
     String message, {
     String? tag,
     dynamic error,
     StackTrace? stacktrace,
   });
 
-  Future<void> error(
+  void error(
     String message, {
     String? tag,
     dynamic error,
-    StackTrace? stacktrace,
+    StackTrace stacktrace,
   });
 }
