+++ layers/log_reporter/lib/src/main.dart
@@ -1,0 +1,6 @@
+import 'package:get_it/get_it.dart';
+import 'package:log_reporter/src/di/di.dart';
+
+Future<void> setUpModule(final GetIt getIt) async {
+  setupPackageDependencies(getIt);
+}
