+++ layers/log_reporter/pubspec.yaml
@@ -1,21 +1,16 @@
 name: log_reporter
-description: A new Flutter package project.
-version: 0.0.1
-homepage:
+description: A Flutter package used for generating logs.
+publish_to: 'none'
 
 environment:
   sdk: '>=3.1.2 <4.0.0'
-  flutter: ">=1.17.0"
 
 dependencies:
-  flutter:
-    sdk: flutter
+  get_it: ^7.6.4
   logger: ^2.0.2+1
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
-  flutter_lints: ^2.0.0
-
-
-flutter:
+  lints: ^3.0.0
+  test: ^1.24.9
+  mocktail: ^0.3.0
+  build_runner: ^2.4.6
